Move Python tests
Prior to this commit, the tests and the debugger were awkward to run
from Pycharm, due to how it added certain directories to the Python path
automatically, and how it interacted with compiled Python extensions.

This moves the tests up one level into a tests directory, which should
simplify the developer experience and everything will just work.